# perf: use `bytes.Index` for substring search