# Description

Here's the benchmarks for KMP and `bytes.Index` (I excluded comparison of heap allocation since both version had zero heap allocations).


```
goos: linux
goarch: amd64
pkg: github.com/ozontech/seq-db/pattern
cpu: 12th Gen Intel(R) Core(TM) i5-12600K
                                              │    kmp.txt    │              bytes.txt               │
                                              │    sec/op     │    sec/op     vs base                │
FindSequence_Deterministic/regular-cases-0-16   19.605n ±  7%   8.456n ±  1%  -56.87% (p=0.000 n=10)
FindSequence_Deterministic/regular-cases-1-16   14.305n ±  2%   5.590n ±  1%  -60.92% (p=0.000 n=10)
FindSequence_Deterministic/corner-cases-0-16     91.35n ±  3%   22.86n ±  1%  -74.98% (p=0.000 n=10)
FindSequence_Deterministic/corner-cases-1-16    167.70n ±  2%   28.44n ±  1%  -83.04% (p=0.000 n=10)
FindSequence_Deterministic/corner-cases-2-16    2482.5n ±  1%   181.8n ±  1%  -92.68% (p=0.000 n=10)
FindSequence_Deterministic/corner-cases-3-16    39.509µ ±  1%   2.663µ ±  1%  -93.26% (p=0.000 n=10)
FindSequence_Random/tiny-16                      63.56n ± 34%   12.10n ± 11%  -80.96% (p=0.000 n=10)
FindSequence_Random/small-16                    190.70n ±  8%   23.70n ± 39%  -87.57% (p=0.000 n=10)
FindSequence_Random/medium-16                   681.70n ± 25%   41.71n ± 28%  -93.88% (p=0.000 n=10)
FindSequence_Random/large-16                    9804.5n ±  2%   538.2n ±  7%  -94.51% (p=0.000 n=10)
FindSequence_Random/extra-large-16              603.54µ ±  1%   58.99µ ±  2%  -90.23% (p=0.000 n=10)
geomean                                          702.6n         94.94n        -86.49%

                                   │    kmp.txt    │                bytes.txt                 │
                                   │      B/s      │      B/s        vs base                  │
FindSequence_Random/tiny-16          960.4Mi ± 52%   5044.3Mi ± 12%   +425.25% (p=0.000 n=10)
FindSequence_Random/small-16         1.250Gi ±  9%   10.424Gi ± 33%   +733.58% (p=0.000 n=10)
FindSequence_Random/medium-16        1.402Gi ± 20%   22.882Gi ± 32%  +1531.58% (p=0.000 n=10)
FindSequence_Random/large-16         1.556Gi ±  2%   28.363Gi ±  7%  +1722.42% (p=0.000 n=10)
FindSequence_Random/extra-large-16   1.618Gi ±  1%   16.555Gi ±  2%   +923.15% (p=0.000 n=10)
geomean                              1.329Gi          14.07Gi         +959.02%
```

So we see that on regular cases we have around 53% of speedup by using `bytes.Index`. And on large strings we see around of 90% speedup (I believe that's because we build LPS for needles, not the haystack).

Here are features of my CPU (it supports both SSE4.2 and AVX2):
```
Flags:                                   fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush dts acpi mmx fxsr sse sse2 ss ht tm pbe syscall nx pdpe1gb rdtscp lm constant_tsc art arch_perfmon pebs bts rep_good nopl xtopology nonstop_tsc cpuid aperfmperf tsc_known_freq pni pclmulqdq dtes64 monitor ds_cpl vmx smx est tm2 ssse3 sdbg fma cx16 xtpr pdcm pcid sse4_1 **sse4_2** x2apic movbe popcnt tsc_deadline_timer aes xsave avx f16c rdrand lahf_lm abm 3dnowprefetch cpuid_fault ssbd ibrs ibpb stibp ibrs_enhanced tpr_shadow flexpriority ept vpid ept_ad fsgsbase tsc_adjust bmi1 **avx2** smep bmi2 erms invpcid rdseed adx smap clflushopt clwb intel_pt sha_ni xsaveopt xsavec xgetbv1 xsaves split_lock_detect user_shstk avx_vnni dtherm ida arat pln pts hwp hwp_notify hwp_act_window hwp_epp hwp_pkg_req hfi vnmi umip pku ospke waitpkg gfni vaes vpclmulqdq rdpid movdiri movdir64b fsrm md_clear serialize pconfig arch_lbr ibt flush_l1d arch_capabilities
```

Closes #98

---

- [x] I have read and followed all requirements in [CONTRIBUTING.md](https://github.com/ozontech/seq-db/blob/main/.github/CONTRIBUTING.md);
- [ ] I used LLM/AI assistance to make this pull request;

If you have used LLM/AI assistance please provide model name and full prompt:

```
Model: {{model-name}}
Prompt: {{prompt}}
```


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- Refactor
  - Switched substring handling to a leaner byte-oriented implementation for multi-part pattern searches, simplifying internals and improving search performance.

- Tests
  - Replaced legacy substring tests with new sequence-focused tests and added deterministic and randomized benchmarks plus data generators to strengthen validation and performance coverage.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->